refactor: change utils dir
