+++ src/utils/getIdFromUrl.ts
@@ -1,0 +1,5 @@
+export const getIdFromUrl = (url: string) => {
+    const urlParts = url.split('/');
+    const idIndex = urlParts.indexOf('users') + 1;
+    return urlParts[idIndex];
+};
